frontend/requirements_simple.txt
streamlit==1.28.1
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.3
requests==2.31.0
plotly==5.18.0
pillow==10.1.0
python-dotenv==1.0.0
